Update navigation.html
Added a span element with an inner span and a img element in it. This section will only get displayed when the screen is in a certain size (the current setting is 480px) or less. When the section is displayed, you'll have a three-line navigation bar with an image of Tux behind it.